Pretty output + Numbered commands = Better Cheats
The file is now read line by line, using 'bufio' package and some sort
of a scanner. Reminds me of Java. Thanks StackOverflow!

Currently the colored output is produced just by using the ANSI escape
codes: http://en.wikipedia.org/wiki/ANSI_escape_code#Colors

But later, something more feature blown can be used:
https://github.com/wsxiaoys/terminal/

Comments are printed in yellow.

Commands are prefixed with a number, that will be used when doing stuff
like:

    cheat git 12 		# Execute the 12th command

    cheat git --copy 12		# Copy 12th command to clipboard

The errors are now printed to stderror, and exit with a status of 1.

When no args are passed, help is shown.